Refactored base's TypeFamily metaclass; type tests
Moved off of python's abc, because I didn't need it, and it was
introducing bugs.  The result is a lot cleaner anyways, in that
there's no need to decorate classes with @typeclass anymore; just
inherit from the metaclass.

Added tests for everything currently in type_defs (except common.py,
which doesn't have much to test atm.) Which means the type code is
officially working.

Noticed that asserts tests are broken, and probably all of
runtime/composition is broken as well. Going to actually remove
runtime's composition and move its functionality into
type_defs/composition.py.