Add a custom function type
New type in type_defs/functions.py:Function, which enforces arity and
return_type in callable instances.

Interestingly, comparing return_type to a different type declaration
here led to some thoughts about whether the current types are alright
for this: some additional __subclasscheck__s may need to be added, but I
think the default should be okay for the most part.

The new Function type purposefully does not interface directly with
schema.  Schema is generally seperate from types and must be, as it
can't be done statically, and the types are needed to work statically --
which honestly with current use of metaclasses probably don't.
The recommended path is to just have Function check for a dict and wrap
the function declaration with @schema normally.