Allow one of multiple possible schema values
This was already accounted for in typecheck, where union types are
trivial.  Schemas were using the same syntax, but are not actually
types and can't actually use the same solution.  This simple class and
handling for it give a possibility for these annotation types.

Example usage may be:

    @schema
    def test(a) -> SchemaOr({"a": Any}, type(None)):
        if a < 0:
          return None
        return {"a": a}

Which would validate given either return statement.