Remove comment saying to provide no-op versions
Originally I was thinking that for production code, the library could
provide a switch that would give you either the real schema and
typecheck or empty functions.  After thinking about this more, there is
not an easily available clean way to do this, and any effort that goes
into that could be spent on more important features.

Dropping in-library switch to provide empty functions instead of schema
and typecheck.